blog: extract post list component (for reuse on home page)
